convert archive config page
